monitoring: generate panel links, link alerts to correct panel
