+++ docker-images/prometheus/cmd/prom-wrapper/receivers.go
@@ -88,7 +88,8 @@ func newRoutesAndReceivers(newAlerts []*schema.ObservabilityAlerts, externalURL 
 
 	// Parameterized alertmanager templates
 	var (
-		dashboardURLTemplate = strings.TrimSuffix(externalURL, "/") + `/-/debug/grafana/d/{{ .CommonLabels.service_name }}/{{ .CommonLabels.service_name }}`
+		// link to grafana dashboard, based on external URL configuration and alert labels
+		dashboardURLTemplate = strings.TrimSuffix(externalURL, "/") + `/-/debug/grafana/d/{{ .CommonLabels.service_name }}/{{ .CommonLabels.service_name }}?viewPanel={{ .CommonLabels.grafana_panel_id }}`
 
 		// messages for different states
 		firingBodyTemplate          = `{{ .CommonLabels.level | title }} alert '{{ .CommonLabels.name }}' is firing for service '{{ .CommonLabels.service_name }}' ({{ .CommonLabels.owner }}).`
