+++ monitoring/monitoring/util.go
@@ -4,9 +4,6 @@ import (
 	"errors"
 	"fmt"
 	"strings"
-	"unicode"
-
-	"github.com/grafana-tools/sdk"
 )
 
 // upperFirst returns s with an uppercase first rune.
@@ -27,25 +24,6 @@ func stringPtr(s string) *string { return &s }
 
 // boolPtr converts a boolean value to a pointer, useful for setting fields in some APIs.
 func boolPtr(b bool) *bool { return &b }
-
-// isValidUID checks if the given string is a valid UID for entry into a Grafana dashboard. This is
-// primarily used in the URL, e.g. /-/debug/grafana/d/syntect-server/<UID> and allows us to have
-// static URLs we can document like:
-//
-// 	Go to https://sourcegraph.example.com/-/debug/grafana/d/syntect-server/syntect-server
-//
-// Instead of having to describe all the steps to navigate there because the UID is random.
-func isValidUID(s string) bool {
-	if s != strings.ToLower(s) {
-		return false
-	}
-	for _, r := range s {
-		if !(unicode.IsLetter(r) || unicode.IsNumber(r) || r == '-') {
-			return false
-		}
-	}
-	return true
-}
 
 // toMarkdown converts a Go string to Markdown, and optionally converts it to a list item if requested by forceList.
 func toMarkdown(m string, forceList bool) (string, error) {
@@ -84,16 +62,4 @@ func toMarkdown(m string, forceList bool) (string, error) {
 	}
 
 	return m, nil
-}
-
-// setPanelSize is a helper to set a panel's size.
-func setPanelSize(p *sdk.Panel, width, height int) {
-	p.GridPos.W = &width
-	p.GridPos.H = &height
-}
-
-// setPanelSize is a helper to set a panel's position.
-func setPanelPos(p *sdk.Panel, x, y int) {
-	p.GridPos.X = &x
-	p.GridPos.Y = &y
 }
